Created Queue Structure with Functions + others
Most of this update is in the queue struct and functions for it. Most of it is very typical of a queue with enequeu, dequeue, peek, isEmpty, etc. I also made some other minor updates such as adding a define for TIME_QUANTUM and creating some prototypes and etc. Added some stuff to the mutex for now but we can focus on that later. Documented all my updates