A bunch of data structures
The majority of what we will need for our data structures. We have a hash table where we will be storing all of our threads. A Priority Linked List sorted based on time_counter values least to greatest. A mlfq that stores an array of size MLFQ_LEVELS of queues. Queues for the MLFQ. Will be preforming some tests before integrating into actual code