Proper Amount of Levels for MLFQ
Changed the code so that there are now 4 levels instead of 5. Tested it and it works when the given inputs are expected. Changed one of the get_mlfq_threads to grab from a nonexisting priority level and the program printed the proper error message and returned NULL as expected. Since the next line tries to print the returned thread, there was a segfault. ALWAYS CHECK IF RETURNED THREADS ARE NULL BEFORE EXECUTING NEXT LINE